Extract directFundingState code into own directory
