+++ packages/wallet/src/redux/direct-funding-store/direct-funding-state/depositing/state.ts
@@ -1,5 +1,5 @@
 import { BaseDirectFundingState, SAFE_TO_DEPOSIT, baseDirectFundingState } from '../state';
-import { TransactionExists } from '../../utils';
+import { TransactionExists } from '../../../utils';
 
 // Deposit status
 export const WAIT_FOR_TRANSACTION_SENT = 'WAIT_FOR_TRANSACTION_SENT';
